code refactor and annotation
